+++ srunner/scenariomanager/carla_data_provider.py
@@ -44,7 +44,7 @@ class CarlaDataProvider(object):
     """
 
     _actor_velocity_map = dict()
-    _actor_location_map = dict()
+    _actor_transform_map = dict()
     _traffic_light_map = dict()
     _map = None
     _world = None
@@ -62,11 +62,11 @@ class CarlaDataProvider(object):
         else:
             CarlaDataProvider._actor_velocity_map[actor] = 0.0
 
-        if actor in CarlaDataProvider._actor_location_map:
+        if actor in CarlaDataProvider._actor_transform_map:
             raise KeyError(
                 "Vehicle '{}' already registered. Cannot register twice!".format(actor.id))
         else:
-            CarlaDataProvider._actor_location_map[actor] = None
+            CarlaDataProvider._actor_transform_map[actor] = None
 
     @staticmethod
     def register_actors(actors):
@@ -85,9 +85,9 @@ class CarlaDataProvider(object):
             if actor is not None and actor.is_alive:
                 CarlaDataProvider._actor_velocity_map[actor] = calculate_velocity(actor)
 
-        for actor in CarlaDataProvider._actor_location_map:
+        for actor in CarlaDataProvider._actor_transform_map:
             if actor is not None and actor.is_alive:
-                CarlaDataProvider._actor_location_map[actor] = actor.get_location()
+                CarlaDataProvider._actor_transform_map[actor] = actor.get_transform()
 
     @staticmethod
     def get_velocity(actor):
@@ -102,16 +102,29 @@ class CarlaDataProvider(object):
             return CarlaDataProvider._actor_velocity_map[actor]
 
     @staticmethod
-    def get_location(actor):
+    def get_transform(actor):
         """
-        returns the location for the given actor
+        returns the transform for the given actor
         """
-        if actor not in CarlaDataProvider._actor_location_map.keys():
+        if actor not in CarlaDataProvider._actor_transform_map.keys():
             # We are initentionally not throwing here
             # This may cause exception loops in py_trees
             return None
         else:
-            return CarlaDataProvider._actor_location_map[actor]
+            return CarlaDataProvider._actor_transform_map[actor]
+
+    @staticmethod
+    def get_location(actor):
+        """
+        returns the transform for the given actor
+        """
+        transform = CarlaDataProvider.get_transform(actor)
+        if transform is None:
+            return None
+            # We are initentionally not throwing here
+            # This may cause exception loops in py_trees
+        else:
+            return transform.location
 
     @staticmethod
     def prepare_map():
@@ -206,7 +219,7 @@ class CarlaDataProvider(object):
         Cleanup and remove all entries from all dictionaries
         """
         CarlaDataProvider._actor_velocity_map.clear()
-        CarlaDataProvider._actor_location_map.clear()
+        CarlaDataProvider._actor_transform_map.clear()
         CarlaDataProvider._traffic_light_map.clear()
         CarlaDataProvider._map = None