+++ srunner/scenarios/basic_scenario.py
@@ -16,7 +16,7 @@ import math
 import carla
 import numpy as np
 
-from srunner.scenariomanager.atomic_scenario_behavior import InTimeToArrivalToLocation
+from srunner.scenariomanager.atomic_scenario_behavior import InTimeToArrivalToTransform
 from srunner.scenariomanager.carla_data_provider import CarlaActorPool, CarlaDataProvider
 from srunner.scenariomanager.scenario_manager import Scenario
 
@@ -65,15 +65,15 @@ class BasicScenario(object):
             criteria = self._create_test_criteria()
 
         # Add a trigger condition for the behavior to ensure the behavior is only activated, when it is relevant
-
-        start_location = None
+        start_transform = None
         if config.trigger_point:
-            start_location = config.trigger_point.location     # start location of the scenario
+            start_transform = config.trigger_point     # start location of the scenario
 
         time_to_start_location = 2.0                               # seconds
         behavior_seq = py_trees.composites.Sequence()
-        if start_location:
-            behavior_seq.add_child(InTimeToArrivalToLocation(self.ego_vehicle, time_to_start_location, start_location))
+        if start_transform:
+            behavior_seq.add_child(InTimeToArrivalToTransform(self.ego_vehicle, time_to_start_location, start_transform))
+
         behavior_seq.add_child(behavior)
 
         self.scenario = Scenario(behavior_seq, criteria, self.name, self.timeout, self.terminate_on_failure)