+++ srunner/scenarios/opposite_vehicle_taking_priority.py
@@ -118,7 +118,7 @@ class OppositeVehicleRunningRedLight(BasicScenario):
         crossing_point_dynamic = get_crossing_point(self.ego_vehicle)
 
         # start condition
-        startcondition = InTriggerDistanceToLocation(
+        startcondition = InTriggerDistanceToTransform(
             self.ego_vehicle,
             crossing_point_dynamic,
             self._ego_distance_to_traffic_light,